+++ rio_tiler/__init__.py
@@ -1,6 +1,6 @@
 """rio-tiler."""
 
-__version__ = "9.0.0a1"
+__version__ = "9.0.0a2"
 
 from . import (  # noqa
     colormap,